Move factor_product and factor_divide to base
