+++ pgmpy/factors/__init__.py
@@ -1,6 +1,9 @@
 from .FactorSet import FactorSet, factorset_product, factorset_divide
+from .base import factor_product, factor_divide
 
 __all__ = ['FactorSet',
            'factorset_divide',
            'factorset_product',
+           'factor_product',
+           'factor_divide'
            ]
