+++ pgmpy/factors/discrete/__init__.py
@@ -1,10 +1,8 @@
-from .DiscreteFactor import State, DiscreteFactor, factor_product, factor_divide
+from .DiscreteFactor import State, DiscreteFactor
 from .CPD import TabularCPD
 from .JointProbabilityDistribution import JointProbabilityDistribution
 
 __all__ = ['TabularCPD',
            'DiscreteFactor',
-           'factor_divide',
-           'factor_product',
            'State'
            ]
