+++ pgmpy/inference/ExactInference.py
@@ -7,7 +7,7 @@ import numpy as np
 from pgmpy.extern.six.moves import filter, range
 
 from pgmpy.extern.six import string_types
-from pgmpy.factors.discrete import factor_product
+from pgmpy.factors import factor_product
 from pgmpy.inference import Inference
 from pgmpy.models import JunctionTree
 from pgmpy.utils import StateNameDecorator
