+++ pgmpy/inference/dbn_inference.py
@@ -2,7 +2,7 @@ from collections import defaultdict
 from itertools import tee, chain, combinations
 
 from pgmpy.factors.discrete import DiscreteFactor
-from pgmpy.factors.discrete import factor_product
+from pgmpy.factors import factor_product
 from pgmpy.inference import Inference, BeliefPropagation
