+++ pgmpy/models/ClusterGraph.py
@@ -5,7 +5,7 @@ from collections import defaultdict
 import numpy as np
 
 from pgmpy.base import UndirectedGraph
-from pgmpy.factors.discrete import factor_product
+from pgmpy.factors import factor_product
 from pgmpy.extern.six.moves import filter, range, zip
