+++ pgmpy/models/FactorGraph.py
@@ -9,7 +9,7 @@ from networkx.algorithms import bipartite
 from pgmpy.models.MarkovModel import MarkovModel
 from pgmpy.base import UndirectedGraph
 from pgmpy.factors.discrete import DiscreteFactor
-from pgmpy.factors.discrete import factor_product
+from pgmpy.factors import factor_product
 from pgmpy.extern.six.moves import filter, range, zip
