+++ pgmpy/models/MarkovModel.py
@@ -6,7 +6,8 @@ import networkx as nx
 import numpy as np
 
 from pgmpy.base import UndirectedGraph
-from pgmpy.factors.discrete import factor_product, DiscreteFactor
+from pgmpy.factors.discrete import DiscreteFactor
+from pgmpy.factors import factor_product
 from pgmpy.independencies import Independencies
 from pgmpy.extern.six.moves import map, range, zip
